refactor: turn service form to modal
