+++ src/types/ui.ts
@@ -1,3 +1,5 @@
+import { Accessor, Setter } from 'solid-js'
+
 export type IncidentsSidebarItem = {
   title: string
   route: string
@@ -6,4 +8,12 @@ export type IncidentsSidebarItem = {
 export type DropdownOption = {
   id: string
   label: string
+}
+
+export type CreateModalProps = {
+  getShouldDisplay: Accessor<Boolean>
+  setShouldDisplay: Setter<Boolean>
+}
+export type ModalProps = {
+  setShouldDisplay: Setter<Boolean>
 }
